catalog_service/requirements.txt
# catalog_service/requirements.txt

uvicorn
fastapi
python-dotenv


